// app/workflow/_components/TaskMenu.tsx
"use client";

import React from "react";
import {
	Accordion,
	AccordionContent,
	AccordionItem,
	AccordionTrigger,
} from "@/components/ui/accordion";
import { TaskType } from "@/types/task";
import { TaskRegistry } from "@/lib/workflow/task/registry";
import { Button } from "@/components/ui/button";
import { Badge } from "@/components/ui/badge";
import { CoinsIcon } from "lucide-react";

export default function TaskMenu() {
	return (
		<aside className="w-[340px] min-w-[300px] max-w-[340px] border-r-2 border-separate h-full p-2 px-4 overflow-auto">
			<Accordion
				type="multiple"
				className="w-full "
				defaultValue={[
					"extraction",
					"interactions",
					"timing",
					"results",
					"storage",
				]}
			>
				<AccordionItem value="interactions">
					<AccordionTrigger className="font-bold">
						User Interactions
					</AccordionTrigger>
					<AccordionContent className="flex flex-col gap-1">
						<TaskMenuBtn
							taskType={TaskType.FILL_INPUT}
						></TaskMenuBtn>
						<TaskMenuBtn
							taskType={TaskType.CLICK_ELEMENT}
						></TaskMenuBtn>
						<TaskMenuBtn
							taskType={TaskType.PRESS_KEY}
						></TaskMenuBtn>
					</AccordionContent>
				</AccordionItem>
				<AccordionItem value="extraction">
					<AccordionTrigger className="font-bold">
						Data Extraction
					</AccordionTrigger>
					<AccordionContent className="flex flex-col gap-1">
						<TaskMenuBtn
							taskType={TaskType.PAGE_TO_HTML}
						></TaskMenuBtn>
						<TaskMenuBtn
							taskType={TaskType.EXTRACT_TEXT_FROM_ELEMENT}
						></TaskMenuBtn>
						<TaskMenuBtn
							taskType={TaskType.EXTRACT_DATA_WITH_AI}
						></TaskMenuBtn>
					</AccordionContent>
				</AccordionItem>
				<AccordionItem value="timing">
					<AccordionTrigger className="font-bold">
						Timing controls
					</AccordionTrigger>
					<AccordionContent className="flex flex-col gap-1">
						<TaskMenuBtn
							taskType={TaskType.WAIT_FOR_ELEMENT}
						></TaskMenuBtn>
					</AccordionContent>
				</AccordionItem>
				<AccordionItem value="storage">
					<AccordionTrigger className="font-bold">
						Data storage
					</AccordionTrigger>
					<AccordionContent className="flex flex-col gap-1">
						<TaskMenuBtn
							taskType={TaskType.READ_PROPERTY_FROM_JSON}
						></TaskMenuBtn>
						<TaskMenuBtn
							taskType={TaskType.ADD_PROPERTY_TO_JSON}
						></TaskMenuBtn>
					</AccordionContent>
				</AccordionItem>
				<AccordionItem value="results">
					<AccordionTrigger className="font-bold">
						Result delivery
					</AccordionTrigger>
					<AccordionContent className="flex flex-col gap-1">
						<TaskMenuBtn
							taskType={TaskType.DELIVER_VIA_WEBHOOK}
						></TaskMenuBtn>
					</AccordionContent>
				</AccordionItem>
			</Accordion>
		</aside>
	);
}

function TaskMenuBtn({ taskType }: { taskType: TaskType }) {
	const task = TaskRegistry[taskType];
	const onDragStart = (event: React.DragEvent, type: TaskType) => {
		event.dataTransfer.setData("application/reactflow", type);
		event.dataTransfer.effectAllowed = "move";
	};
	return (
		<Button
			variant={"secondary"}
			className="flex justify-between items-center gap-2 border w-full"
			draggable
			onDragStart={(event) => onDragStart(event, taskType)}
		>
			<div className="flex gap-2">
				<task.icon size={20} />
				{task.label}
			</div>
			<Badge className="gap-2 flex items-center " variant={"outline"}>
				<CoinsIcon size={16} />
				{task.credits}
			</Badge>
		</Button>
	);
}
